# pos-api/target/classes/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Products
  license:
    name: MIT
paths:
...
  /items:
    get:
      summary: List all items
      operationId: listItems
      tags:
        - items
      responses:
        '200':
          description: A paged array of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Items"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - item
      operationId: addItem
      summary: Create a item
      description: Creates a item.
      requestBody:
        description: The item
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemFields'
        required: true
      responses:
        201:
          description: Item created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /items/{itemId}:
    get:
      summary: Info for a specific item
      operationId: showItemById
      tags:
        - item
      parameters:
        - name: itemId
          in: path
          required: true
          description: The id of the item to retrieve
          schema:
            type: long
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - item
      operationId: updateItem
      summary: Update a item's details
      description: Updates the item record with the specified details.
      parameters:
        - name: itemId
          in: path
          description: The ID of the item.
          required: true
          schema:
            type: long
      requestBody:
        description: The item details to use for the update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemFields'
        required: true
      responses:
        200:
          description: Item details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - item
      operationId: deleteItem
      summary: Delete an item by ID
      description: Returns the item or a 404 error.
      parameters:
        - name: itemId
          in: path
          description: The ID of the item.
          required: true
          schema:
            type: long
      responses:
        200:
          description: Item details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Owner  not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /items/{itemId}/addone:
    put:
      tags:
        - item
      operationId: addItemById
      parameters:
        - name: itemId
          in: path
          description: The ID of the item.
          required: true
          schema:
            type: long
      responses:
        200:
          description: Item details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /items/{itemId}/subone:
    put:
      tags:
        - item
      operationId: subItemById
      parameters:
        - name: itemId
          in: path
          description: The ID of the item.
          required: true
          schema:
            type: long
      responses:
        200:
          description: Item details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /orders:
    get:
      summary: List all orders
      operationId: listOrders
      tags:
        - orders
      responses:
        '200':
          description: A paged array of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - order
      operationId: addOrder
      summary: Create a order
      description: Creates a order.
      requestBody:
        description: The order
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderFields'
        required: true
      responses:
        201:
          description: order created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /orders/generate:
    post:
      tags:
        - order
      operationId: generateOrder
      summary: Create a order
      description: Creates a order.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderFields"
        required: true
      responses:
        201:
          description: order created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /orders/{orderId}:
    get:
      summary: Info for a specific order
      operationId: showOrderById
      tags:
        - order
      parameters:
        - name: orderId
          in: path
          required: true
          description: The id of the order to retrieve
          schema:
            type: long
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    UserFields:
      type: object
      properties:
        name:
          title: name
          description: The name of the user.
          type: string
          minLength: 1
          maxLength: 30
          pattern: '^[a-zA-Z]*$'
          example: George
        email:
          title: Email
          description: The email of the user.
          type: string
          minLength: 1
          maxLength: 255
          example: 'abc@ef.com'
        pass:
          title: Password
          description: The password of the user.
          type: string
          format: password
          minLength: 1
          maxLength: 80
          example: '123456789'
        money:
          title: Money
          description: The money of the user.
          type: double
          minimum: 0
        address1:
          type: string
          minLength: 1
          maxLength: 20
          pattern: '^[0-9]*$'
          example: '10086'
        address2:
          type: string
          minLength: 1
          maxLength: 20
          pattern: '^[0-9]*$'
          example: '10087'
        contact:
          type: string
          pattern: '^\d{0,20}$'
        symbol:
          type: string
        footer:
          type: string
        image:
          type: string
      required:
        - name
    TaxFields:
      type: object
      properties:
        tax:
          type: boolean
        percentage:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
      required:
        - tax
    User:
      title: User
      allOf:
        - $ref: '#/components/schemas/UserFields'
        - $ref: '#/components/schemas/TaxFields'
        - type: object
          properties:
            id:
              type: long
            cartId:
              type: long
          required:
            - id
    ProductFields:
      type: object
      properties:
        pid:
          type: string
        name:
          type: string
        price:
          type: double
          minimum: 0
        image:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        quantity:
          type: integer
          format: int32
          minimum: 0
      required:
        - name
    Product:
      allOf:
        - $ref: '#/components/schemas/ProductFields'
        - type: object
          properties:
            id:
              type: long
            stock:
              type: boolean
          required:
            - id
    ItemFields:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
          format: int32
          minimum: 0
      required:
        - product
        - quantity
    Item:
      allOf:
        - $ref: '#/components/schemas/ItemFields'
        - type: object
          properties:
            id:
              type: long
            cartId:
              type: long
          required:
            - id
    CartFields:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
      required:
        - items
    Cart:
      allOf:
        - $ref: '#/components/schemas/CartFields'
        - type: object
          properties:
            id:
              type: long
          required:
            - id
    CategoryFields:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Category:
      allOf:
        - $ref: '#/components/schemas/CategoryFields'
        - type: object
          properties:
            id:
              type: long
          required:
            - id
    OrderFields:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        cart:
          $ref: '#/components/schemas/Cart'
      required:
        - user
        - cart
    Order:
      allOf:
        - $ref: '#/components/schemas/OrderFields'
        - type: object
          properties:
            id:
              type: long
          required:
            - id
...
    Items:
      type: array
      items:
        $ref: "#/components/schemas/Item"
...
    Orders:
      type: array
      items:
        $ref: "#/components/schemas/Order"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string